Introduce JobExecution and JobExecutor
**Job Execution**

Job.run starts a new job execution, but it returned a
`Future[JobStartStatus]` (which should be completed once the
job was started). Once the job was finished, it had to invoke
`JobContext.finishedCallback()`.
This was not intuitive and error prone, it was easy to forget
to invoke the `finishedCallback`.

Additionally `cancel()` was defined on the Job, so that the Job
singleton had to e.g. keep a `cancel` flag and evaluate this
inside `run()`. This cancel flag then had to be reset at the
beginning of `run()`, otherwise a new execution would start
with the cancel flag still being set.

By introduction the `JobExecution` which is returned by
`Job.run()`, these potential issues are eliminated. A job
execution is considered to be completed once the
`JobExecution.result` future is completed. Also `cancel()`
is now defined by the `JobExecution`.

**Job Executor**

Jobs were previously directly `run()` by the `JobManager`, but
it knew nothing about currently executed jobs (this was not
possible because job execution was hidden and not modeled
explicitely).

Now the `JobManager` delegates job execution to the
`JobExecutor`, which also keeps a list of currently executed
jobs.

The `JobContext` now also holds the jobType, which was useful
to have in the `JobExecutor`. It also makes `ActorJob`s more
easy because they no longer have to provide the `def jobType`.