C* optimization: avoid * queries for jobs and locks.
"*-queries" without any restriction on the partition key are
very expensive and should be avoided if possible.
Therefore such queries for jobs and locks are now done for each
JobType/LockType (in parallel). In consequence, the LockTypes and
JobTypes were changed from a trait to a class that expect a list
of all known values, not just a PartialFunction for a by name
lookup. If the companion objects are used, this change is source
and binary compatible, therefore we don't increment the major version
for this (even if it would be required if we'd follow semver
strictly).

This also fixes issue #6