Really Introduce a default limitByJobType val in JobStatusRepository
This commit introduces a new defaultLimitByJobType function that is used as default value in
getMetadata to limit the number of returned job status entries.

This is necessary to have a stable reference to a function that that satisfies the Mockito verify check
in JobUpdaterSpec and emulate the old defaultValue behaviour of getMetadata in JobStatusRepository.